Refactor clientService into it's own module. Refactor websocket command and portlist message handlers into separate functions.
